Общее улучшение структуры, упорядочивание файлов по слоям/пакетам исходя из их назначения
- cli - точка входа в приложение, графический клиент
- domain - корневая логика приложения, учёт сценариев
- music_api - клиентский инфраструктурный слой, дёргает апи, используя клиент
