+++ ymd/__init__.py
@@ -1,0 +1,10 @@
+'''
+    Приложение состоит из 3 общих слоёв:
+
+    CLI (точка входа, интерфейс)
+    |
+    Domain (бизнес-логика, обработка данных, подготовка к скачиванию, учёт запроса пользователя)
+    |
+    API (транспортный слой, клиентский код для работы с Yandex Music API, возвращает данные)
+    
+'''
