+++ ymd/domain/__init__.py
@@ -1,0 +1,3 @@
+'''
+    Модуль, содержащий основную бизнес-логику приложения
+'''
