+++ ymd/music_api/__init__.py
@@ -3,4 +3,6 @@
 
     Содержит модели для сериализации и десериализации данных,
     а также функции для получения информации о треках и их скачивания.
+
+    Фактически модуль содержит клиентский код для работы с Yandex Music API.
 '''
